Fix build and remove warnings
Adapt GetDeliveryPath and GetDeliveryIniterary to return the same thing

+++ Projects/Project1/src/order.h
@@ -4,15 +4,15 @@
 class Order
 {
 public:
-    Order(double weight, uint volume) : _weight(weight), _volume(volume) { }
+    Order(uint weight, uint volume) : _weight(weight), _volume(volume) { }
     Order(const Order& other) : _weight(other._weight), _volume(other._volume) { }
     Order operator=(const Order& other) const { return Order(other); }
 
-    double    GetWeight() const { return _weight; }
+    uint      GetWeight() const { return _weight; }
     uint      GetVolume() const { return _volume; }
 
 private:
-    const double    _weight;
+    const uint      _weight;
     const uint      _volume;
 };
